pyproject.toml
[project]
name = "local_rag"
version = "0.1.0"
description = "Resilient multi-strategy RAG evaluation architecture"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "llama-index>=0.14.20",
    "llama-index-graph-stores-ladybug>=0.3.0",
    "llama-cpp-python>=0.3.0",
    "marimo>=0.9.0",
    "openai>=2.31.0",
    "pydantic-ai-slim>=1.78.0",
    "real-ladybug>=0.15.3",
    "sqlite-vec>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/db", "src/evaluation", "src/grading", "src/ingestion", "src/models"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ty>=0.0.29",
]
